+++ spacypdfreader/parsers/pdfminer.py
@@ -10,17 +10,17 @@ def parser(pdf_path: str, page_number: int, **kwargs):
 
     Parameters:
         pdf_path: Path to a PDF file.
-        page_number: The page number of the PDF to convert from PDF to text. Must be one 
-            digit based indexing (e.g. the first page of the PDF is page 1, as 
+        page_number: The page number of the PDF to convert from PDF to text. Must be one
+            digit based indexing (e.g. the first page of the PDF is page 1, as
             opposed to page 0).
-        **kwargs: `**kwargs` will be passed to 
+        **kwargs: `**kwargs` will be passed to
             [`pdfminer.high_level.extract_text`](https://pdfminersix.readthedocs.io/en/latest/reference/highlevel.html#extract-text).
 
     Returns:
         str: The PDF page as a string.
 
     Examples:
-        pdfminer is the default PDF to text parser and will be automatically 
+        pdfminer is the default PDF to text parser and will be automatically
         used unless otherwise specified.
 
         >>> import spacy
@@ -63,5 +63,6 @@ def parser(pdf_path: str, page_number: int, **kwargs):
 
 class PdfminerParser:
     """This class has bee included for backwards compatibility. Do not use."""
+
     def __init__(self):
         return None