+++ tests/test_pytesseract.py
@@ -1,6 +1,5 @@
 import spacy
-from spacypdfreader import __version__
-from spacypdfreader._utils import _get_number_of_pages
+
 from spacypdfreader.parsers import pytesseract
 from spacypdfreader.spacypdfreader import pdf_reader
 
@@ -34,19 +33,25 @@ def test_pytesseract_with_params():
 
 def test_pytesseract_multi():
     nlp = spacy.load("en_core_web_sm")
-    doc = pdf_reader("tests/data/test_pdf_01.pdf", nlp, pytesseract.parser, n_processes=4)
+    doc = pdf_reader(
+        "tests/data/test_pdf_01.pdf", nlp, pytesseract.parser, n_processes=4
+    )
     pdf_assertions(doc)
 
 
 def test_pytesseract_multi_with_params():
     params = {"nice": 1}
     nlp = spacy.load("en_core_web_sm")
-    doc = pdf_reader("tests/data/test_pdf_01.pdf", nlp, pytesseract.parser, n_processes=4, **params)
+    doc = pdf_reader(
+        "tests/data/test_pdf_01.pdf", nlp, pytesseract.parser, n_processes=4, **params
+    )
     pdf_assertions(doc)
 
 
 def test_pytesseract_multi_same_as_single():
     nlp = spacy.load("en_core_web_sm")
-    doc_multi = pdf_reader("tests/data/test_pdf_01.pdf", nlp, pytesseract.parser, n_processes=4)
+    doc_multi = pdf_reader(
+        "tests/data/test_pdf_01.pdf", nlp, pytesseract.parser, n_processes=4
+    )
     doc_single = pdf_reader("tests/data/test_pdf_01.pdf", nlp, pytesseract.parser)
     assert doc_multi.text == doc_single.text